:recycle: refactor: Refatoração feita no Core

+++ Auction.Core/Models/Bidder.cs
@@ -2,6 +2,6 @@
 
 public class Bidder
 {
-  public int Id { get; set;}
-  public string Name { get; set;} = string.Empty;
+  public long Id { get; set; }
+  public string Name { get; set; } = string.Empty;
 }
